Add abstract address class
